Complete 4.21
Add user token check before deleting blogs

+++ part4/bloglist/controllers/blogs.js
@@ -32,6 +32,7 @@ blogsRouter.put('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
+  //User token authorization
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid ' })
@@ -61,12 +62,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
+  //User token authorization
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid ' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const deletedBlog = await Blog.findById(request.params.id)
   if (deletedBlog === null) {
-    response.status(404).json({ error: 'blog not found' })
-  } else {
+    return response.status(404).json({ error: 'blog not found' })
+  } else if (deletedBlog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   }
+
 })
 
 module.exports = blogsRouter
